requirements.txt
Flask
Flask-Bcrypt
Flask-SQLAlchemy
python-dotenv
spotipy
TextBlob
SpeechRecognition
opencv-python
fer
gunicorn
